refactor(orc8r): Split health to protected and southbound
