+++ feg/cloud/go/services/health/reporter/network_health_reporter.go
@@ -25,7 +25,7 @@ import (
 	"magma/feg/cloud/go/serdes"
 	"magma/feg/cloud/go/services/health"
 	"magma/feg/cloud/go/services/health/metrics"
-	"magma/feg/cloud/go/services/health/servicers"
+	servicers "magma/feg/cloud/go/services/health/servicers/southbound"
 	"magma/orc8r/cloud/go/orc8r"
 	"magma/orc8r/cloud/go/services/configurator"
 )
