+++ feg/cloud/go/services/health/servicers/southbound/config.go
@@ -32,6 +32,13 @@ const (
 
 var defaultServices = []string{"SWX_PROXY", "SESSION_PROXY"}
 
+type healthConfig struct {
+	services              []string
+	cpuUtilThreshold      float32
+	memAvailableThreshold float32
+	staleUpdateThreshold  uint32
+}
+
 func GetHealthConfigForNetwork(ctx context.Context, networkID string) *healthConfig {
 	defaultConfig := &healthConfig{
 		services:              defaultServices,
